android/handsfree-client: Fix index type in call action command
According to BT HAL API index should be int.

+++ android/hal-msg.h
@@ -1224,7 +1224,7 @@ struct hal_cmd_hf_client_dial_memory {
 #define HAL_OP_HF_CLIENT_CALL_ACTION		0x0a
 struct hal_cmd_hf_client_call_action {
 	uint8_t action;
-	uint8_t index;
+	int32_t index;
 } __attribute__((packed));
 
 #define HAL_OP_HF_CLIENT_QUERY_CURRENT_CALLS	0x0b
